+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# cranko 0.0.5 (2020-08-23)
+
+- Add `github install-credential-helper` and hidden support command `github
+  _credential-helper`; this release will help check whether they actually work!
+
 # cranko 0.0.4 (2020-08-23)
 
 - Fix up the analysis of release histories to work properly when projects are not